requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.9
djangorestframework==3.12.4
jsonschema==4.2.1
pyrsistent==0.18.0
pytz==2021.3
rfc3987==1.3.8
sqlparse==0.4.2
